backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
pymupdf==1.24.5
groq==0.9.0
langgraph==0.1.19
langchain-core==0.2.23
motor==3.5.0
redis==5.0.7
pydantic==2.8.2
python-multipart==0.0.9
sentence-transformers==3.0.1
numpy==1.26.4
